+++ visualization/app/codeCharta/state/store/appStatus/appStatus.reducer.ts
@@ -1,11 +1,11 @@
 import { combineReducers } from "redux"
 
-import { hoveredBuildingPath } from "./hoveredBuildingPath/hoveredBuildingPath.reducer"
+import { hoveredNodeId } from "./hoveredNodeId/hoveredNodeId.reducer"
 import { rightClickedNodeData } from "./rightClickedNodeData/rightClickedNodeData.reducer"
 import { selectedBuildingId } from "./selectedBuildingId/selectedBuildingId.reducer"
 
 const appStatus = combineReducers({
-	hoveredBuildingPath,
+	hoveredNodeId,
 	selectedBuildingId,
 	rightClickedNodeData
 })