+++ visualization/app/codeCharta/ui/attributeSideBar/attributeTypeSelector/attributeTypeSelector.module.ts
@@ -3,11 +3,10 @@ import { CommonModule } from "@angular/common"
 
 import { AttributeTypeSelectorComponent } from "./attributeTypeSelector.component"
 import { MaterialModule } from "../../../../material/material.module"
-import { AggregationTypePipe } from "./aggregationType.pipe"
 
 @NgModule({
 	imports: [CommonModule, MaterialModule],
-	declarations: [AttributeTypeSelectorComponent, AggregationTypePipe],
+	declarations: [AttributeTypeSelectorComponent],
 	exports: [AttributeTypeSelectorComponent]
 })
 export class AttributeTypeSelectorModule {}