+++ visualization/app/codeCharta/ui/metricType/metricType.module.ts
@@ -1,5 +1,0 @@
-import "../../state/state.module"
-import angular from "angular"
-import { metricTypeComponent } from "./metricType.component"
-
-angular.module("app.codeCharta.ui.metricType", ["app.codeCharta.state"]).component(metricTypeComponent.selector, metricTypeComponent)